refactor(libs): isolate awards presentation code
