+++ apps/shell-e2e/src/unit/module-boundaries.spec.ts
@@ -37,8 +37,9 @@ describe("remote module boundaries", () => {
       import "@site/layout";
       import "../../../software/src/page";
       import "../../../bio/src/page";
+      import "../../../awards/src/award-emblem";
     `);
-    expect(messages).toHaveLength(3);
+    expect(messages).toHaveLength(4);
     expect(messages.every((message) => message.severity === 2)).toBe(true);
   }, 15_000);
 });
