--- libs/analytics/project.json
@@ -1,27 +1,0 @@
-{
-  "name": "analytics",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/analytics/src",
-  "projectType": "library",
-  "tags": ["type:shared", "scope:analytics"],
-  "targets": {
-    "lint": {
-      "executor": "nx:run-commands",
-      "options": { "command": "biome lint libs/analytics" }
-    },
-    "typecheck": {
-      "executor": "nx:run-commands",
-      "options": { "command": "tsc -p libs/analytics/tsconfig.json" }
-    },
-    "test": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "vitest run --config libs/analytics/vite.config.ts --coverage"
-      }
-    },
-    "build": {
-      "executor": "nx:run-commands",
-      "options": { "command": "tsc -p libs/analytics/tsconfig.json" }
-    }
-  }
-}
