--- libs/analytics/src/index.ts
@@ -1,1 +1,0 @@
-export { type AnalyticsEvent, trackEvent } from "./gtag";
