--- libs/analytics/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": { "types": ["vitest/globals"] },
-  "include": ["src/**/*.ts", "vite.config.ts"]
-}
