+++ libs/design-system/src/index.ts
@@ -1,3 +1,1 @@
 import "./theme.css";
-
-export { AwardEmblem, type AwardIcon } from "./award-emblem";
