--- libs/ui/project.json
@@ -1,25 +1,0 @@
-{
-  "name": "ui",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/ui/src",
-  "projectType": "library",
-  "tags": ["type:shared", "scope:ui"],
-  "targets": {
-    "lint": {
-      "executor": "nx:run-commands",
-      "options": { "command": "biome lint libs/ui" }
-    },
-    "typecheck": {
-      "executor": "nx:run-commands",
-      "options": { "command": "tsc -p libs/ui/tsconfig.json" }
-    },
-    "test": {
-      "executor": "nx:run-commands",
-      "options": { "command": "vitest run --passWithNoTests libs/ui" }
-    },
-    "build": {
-      "executor": "nx:run-commands",
-      "options": { "command": "tsc -p libs/ui/tsconfig.json" }
-    }
-  }
-}
