--- libs/ui/src/index.ts
@@ -1,1 +1,0 @@
-export type Tone = "default" | "muted";
